src/django/requirements.txt
Django==3.1.1
django-environ==0.4.5
django-cors-headers==3.5.0
djangorestframework==3.11.1
djangorestframework-simplejwt==4.6.0
djangorestframework-filters==1.0.0.dev0
django-extensions==3.1.3
drf-spectacular==0.14.0
gunicorn==20.0.4
pre-commit==2.7.1
psycopg2-binary==2.8.6
